[PSL-112] Implement Action Activation Ticket (#87)
* [PSL-112] Implement Action Activation Ticket

Implemented new ticket type: Action Activation (action-act):
 - this ticket can be created by any Pastel node
 - this ticket can only be created if action-reg (Action Registration ticket) that it references exists
 - the transaction with this ticket includes ticket payments to SNs
 Action Activation Ticket json:
 {
    "ticket": {
        "type": "action-act",
        "version": integer,      // version of the blockchain representation of ticket, 1 now
        "pastelID": string,      // PastelID of the caller
        "reg_txid": string,      // txid of the action registration ticket
        "called_at": int,        // block at which action was called (`action_ticket` was created)
                                 // is used to check if the MN that created action registration ticket was indeed the top MN when the the action was called
        "storage_fee": int,      // should match the storage fee from the Action Ticket
        "signature": bytes       // Signature of the ticket created using the MasterNode's private key
    }
}

RPC API:
  - implemented action ticket activation rpc API:
    tickets activate action "reg-ticket-txid" "called-at-height" "fee" "PastelID" "passphrase" ["address"]
  - implemented "find action-act "key") rcp API to find action activation tickets:
     - by ActionReg ticket txid
	 - by Action Caller's Pastel ID
	 - by Action Called-At height (block height at what original Action registration ticket was created)
  - implemented "tickets list action-act ("filter") ("minheight")" rpc API to list action activation tickets
  - implemented filters for "tickets list action ("filter")" rpc API:
       - all: lists all Action tickets (including non-confirmed). Default.
	   - active: lists only activated Action tickets - with ActionAct ticket.
	   - inactive: lists only non-activated Active tickets - without ActionAct ticket created (confirmed).
  - ticket activation now can be called using two different APIs (calls the same code):
       NFT: tickets activate nft ...
	        tickets register act ...
    Action: tickets activate action ...
	        tickets register action-act ...

Refactoring:
  - mnode-rpc.cpp - moved to separate files:
    - "ingest" rpc API -> mnode/rpc/ingest.cpp|h
	- -> mnode/rpc/masterbroadcast.cpp|h
	- "tickets activate" rpc API -> mnode/rpc/tickets-activate.cpp|h
  - fixed some warnings in addrman.cpp, hash.cpp, net.cpp, pow.cpp, blockchain.cpp, httprpc.cpp, mnode-controller.cpp
  - used R-strings in mining rpc API: rpc/mining.cpp

Tests:
  - moved some tests from bitcoin_test to pastel_gtest:
     addrman, base32,base58,base64,bech32,checkpoints,hash,miner,sanity,timedata.
  - python tests -> mn_tickets.py:
     - implemented tests for action activation ticket
	 - refactored nft-act,royalty tests to use assert_raises_rpc API

* [PSL-112] added missing gtest/data files

* - added common class CPastelTicketMNFee for both NFT & Action Activation tickets

 This class is used to calculate MN fees depending on:
   - all MN fees percentage
   - principal MN fee share percentage
   - other MNs fee share percentage
- changed NFT & Action Activation ticket classes to use different MN fees
- corrected MN fees validation for activation tickets in ticket processor
- changed MN fees validation in python tests mn_tickets.py

+++ src/rpc/blockchain.cpp
@@ -690,8 +690,8 @@ Examples:
 
     LOCK(cs_main);
 
-    int nCheckLevel = GetArg("-checklevel", 3);
-    int nCheckDepth = GetArg("-checkblocks", 288);
+    int nCheckLevel = static_cast<int>(GetArg("-checklevel", 3));
+    int nCheckDepth = static_cast<int>(GetArg("-checkblocks", 288));
     if (params.size() > 0)
         nCheckLevel = params[0].get_int();
     if (params.size() > 1)
